delete cast

+++ firmware/controllers/can/can_verbose.cpp
@@ -72,7 +72,7 @@ struct Status{
 
 static void populateFrame(Status& msg) {
   msg.warningCounter = static_cast<uint16_t>(engine->engineState.warnings.lastErrorCode);
-  msg.lastErrorCode = static_cast<uint16_t>(engine->engineState.warnings.recentWarnings.getCount()); //getNextErrorCode() ;
+  msg.lastErrorCode = engine->engineState.warnings.recentWarnings.getCount(); //getNextErrorCode() ;
 	msg.revLimit = !engine->module<LimpManager>()->allowInjection() || !engine->module<LimpManager>()->allowIgnition();
 	msg.mainRelay = enginePins.mainRelay.getLogicValue();
 	msg.fuelPump = enginePins.fuelPumpRelay.getLogicValue();
